+++ utility/test/module_test.go
@@ -1,6 +1,7 @@
 package test
 
 import (
+	"encoding/hex"
 	"math/big"
 	"os"
 	"testing"
@@ -30,6 +31,7 @@ var (
 	defaultNonceString         = utilTypes.BigIntToString(defaults.DefaultAccountAmount)
 	defaultSendAmountString    = utilTypes.BigIntToString(defaultSendAmount)
 	testSchema                 = "test_schema"
+	testMessageSendType        = "MessageSend"
 )
 
 var persistenceDbUrl string
@@ -97,4 +99,20 @@ func newTestPersistenceModule(t *testing.T, databaseUrl string) modules.Persiste
 	require.NoError(t, err)
 
 	return persistenceMod.(modules.PersistenceModule)
+}
+
+func requireValidTestingTxResults(t *testing.T, tx *utilTypes.Transaction, txResults []modules.TxResult) {
+	for _, txResult := range txResults {
+		msg, err := tx.GetMessage()
+		sendMsg, ok := msg.(*utilTypes.MessageSend)
+		require.True(t, ok)
+		require.NoError(t, err)
+		require.Equal(t, int32(0), txResult.GetResultCode())
+		require.Equal(t, "", txResult.GetError())
+		require.Equal(t, testMessageSendType, txResult.GetMessageType())
+		require.Equal(t, int32(0), txResult.GetIndex())
+		require.Equal(t, int64(0), txResult.GetHeight())
+		require.Equal(t, hex.EncodeToString(sendMsg.ToAddress), txResult.GetRecipientAddr())
+		require.Equal(t, hex.EncodeToString(sendMsg.FromAddress), txResult.GetSignerAddr())
+	}
 }