+++ utility/doc/CHANGELOG.md
@@ -7,7 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.0.0.7] - 2022-10-12
+## [0.0.0.7] - 2022-10-14
+
+- Added session interfaces and diagrams
+- Moved `TxIndexer` package to persistence module
+- Added new proto structure `DefaultTxResult`
+- Integrated the `TxIndexer` into the lifecycle
+  - Captured `TxResult` from each played transaction
+  - Moved the storage of transactions to the Consensus Module
+  - Returned the `TxResults` in the `ApplyBlock()` and `GetProposalTransactions()`
+  - `AnteHandleMessage()` now returns `signer`
+  - `ApplyTransaction()` returns `TxResult`
 
 ### [#235](https://github.com/pokt-network/pocket/pull/235) Config and genesis handling